# web/default/src/features/docs/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexus Reach Seedance 2.0 API",
    "version": "2026-07-17",
    "description": "Seedance 2.0 国内视频生成与素材资产接口。所有公开请求统一通过 Nexus Reach API Gateway 访问，并使用 Nexus Reach API Key 鉴权。"
  },
  "servers": [
    {
      "url": "https://gateway.nexus-reach.com",
      "description": "Nexus Reach API Gateway"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/video/generations": {
      "post": {
        "tags": ["视频生成"],
        "summary": "创建 Seedance 2.0 视频任务",
        "operationId": "createSeedanceDomesticVideo",
        "description": "支持文本、图片、视频和音频组合输入。`content` 中最多包含 9 张图片、3 个视频和 3 个音频；音频不能单独作为输入。使用素材库资源时，将 `asset://<AssetId>` 写入对应媒体对象的 `url` 字段。提示词应按媒体在 `content` 中出现的顺序使用“图片1、图片2、视频1”等称呼，禁止直接在提示词中填写 AssetId。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoGenerationRequest"
              },
              "example": {
                "model": "doubao-seedance-2-0-260128",
                "content": [
                  {
                    "type": "image_url",
                    "image_url": {
                      "url": "asset://asset-20260716-reference"
                    },
                    "role": "reference_image"
                  },
                  {
                    "type": "text",
                    "text": "图片1中的虚拟人物在海边自然行走，镜头缓慢向前推进。"
                  }
                ],
                "audio_status": 1,
                "resolution": "720p",
                "ratio": "16:9",
                "dur": 5
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "任务创建成功。请保存返回的公开任务 ID，用于后续查询。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoSubmitResponse"
                },
                "example": {
                  "id": "task_01jzseedanceexample",
                  "task_id": "task_01jzseedanceexample",
                  "object": "video",
                  "model": "doubao-seedance-2-0-260128",
                  "status": "queued",
                  "progress": 0,
                  "created_at": 1784169600
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或内容未通过校验。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "message": "dur must be -1 or an integer between 4 and 15",
                    "type": "invalid_request_error",
                    "code": "invalid_request"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "502": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Nexus Reach API Key，例如：`Authorization: Bearer sk-your-key`。"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "未提供 API Key，或 API Key 无效。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "invalid API key",
                "type": "authentication_error",
                "code": "invalid_api_key"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "当前 API Key 无权访问该接口或资源。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "permission denied",
                "type": "permission_error",
                "code": "permission_denied"
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "请求频率超过限制，请稍后重试。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "rate limit exceeded",
                "type": "rate_limit_error",
                "code": "rate_limit_exceeded"
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "服务内部错误。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "internal server error",
                "type": "server_error",
                "code": "internal_error"
              }
            }
          }
        }
      },
      "UpstreamError": {
        "description": "生成或素材服务暂时不可用，请稍后重试。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "service temporarily unavailable",
                "type": "server_error",
                "code": "service_unavailable"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "ContentMedia": {
        "type": "object",
        "description": "媒体对象。url 可以是符合要求的公网 URL，也可以是已处理为 Active 的素材库地址 `asset://<AssetId>`。",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "description": "媒体地址。"
          }
        },
        "required": ["url"],
        "additionalProperties": false
      },
      "ContentItem": {
        "type": "object",
        "description": "输入内容项。type 与对应内容字段必须匹配。媒体的编号按其在 content 数组中的同类型出现顺序计算。",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["text", "image_url", "video_url", "audio_url"],
            "description": "内容类型。"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "description": "type=text 时必填。引用媒体时请使用“图片1、图片2、视频1”等顺序称呼，不要填写 AssetId。"
          },
          "image_url": {
            "$ref": "#/components/schemas/ContentMedia"
          },
          "video_url": {
            "$ref": "#/components/schemas/ContentMedia"
          },
          "audio_url": {
            "$ref": "#/components/schemas/ContentMedia"
          },
          "role": {
            "type": "string",
            "enum": [
              "reference_image",
              "first_frame",
              "reference_video",
              "reference_audio"
            ],
            "description": "媒体语义角色；仅在相应输入模式需要时填写。"
          }
        },
        "oneOf": [
          {
            "required": ["type", "text"],
            "properties": { "type": { "enum": ["text"] } }
          },
          {
            "required": ["type", "image_url"],
            "properties": { "type": { "enum": ["image_url"] } }
          },
          {
            "required": ["type", "video_url"],
            "properties": { "type": { "enum": ["video_url"] } }
          },
          {
            "required": ["type", "audio_url"],
            "properties": { "type": { "enum": ["audio_url"] } }
          }
        ]
      },
      "VideoGenerationRequest": {
        "type": "object",
        "description": "必须提供 content，或使用 prompt/image/images 简写字段提供至少一种文本、图片或视频输入；音频不能作为唯一输入。",
        "properties": {
          "model": {
            "type": "string",
            "enum": ["doubao-seedance-2-0-260128"],
            "default": "doubao-seedance-2-0-260128",
            "description": "模型名称。"
          },
          "content": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/ContentItem"
            },
            "description": "主要输入内容。最多 9 张图片、3 个视频、3 个音频；音频不能单独作为输入。"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "description": "提示词简写。引用媒体时使用“图片1、视频1”等顺序称呼，不要直接填写 AssetId。"
          },
          "image": {
            "type": "string",
            "minLength": 1,
            "description": "单张图片地址简写。"
          },
          "images": {
            "type": "array",
            "minItems": 1,
            "maxItems": 9,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "minLength": 1
            },
            "description": "多张图片地址简写，最多 9 张。"
          },
          "audio_status": {
            "type": "integer",
            "enum": [0, 1],
            "default": 1,
            "description": "1 自动生成匹配人声、音效和背景音乐；0 生成无声视频。有声视频统一为单声道。"
          },
          "generate_audio": {
            "type": "boolean",
            "description": "audio_status 的布尔别名。"
          },
          "resolution": {
            "type": "string",
            "enum": ["720p", "1080p"],
            "default": "720p",
            "description": "输出分辨率。4K 需渠道单独开通，且尚未完成端到端验证，因此当前不作为公开可用枚举。"
          },
          "ratio": {
            "type": "string",
            "enum": ["adaptive", "16:9", "4:3", "1:1", "3:4", "9:16", "21:9"],
            "default": "adaptive",
            "description": "输出宽高比。"
          },
          "dur": {
            "type": "integer",
            "default": 5,
            "description": "输出秒数。支持 4～15 的整数；-1 表示由模型在 4～15 秒内决定。",
            "oneOf": [{ "enum": [-1] }, { "minimum": 4, "maximum": 15 }]
          }
        },
        "required": ["model"],
        "anyOf": [
          { "required": ["content"] },
          { "required": ["prompt"] },
          { "required": ["image"] },
          { "required": ["images"] }
        ]
      },
      "VideoSubmitResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "公开任务 ID。"
          },
          "task_id": {
            "type": "string",
            "description": "兼容任务 ID 字段，与 id 相同。"
          },
          "object": {
            "type": "string",
            "enum": ["video"]
          },
          "model": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "in_progress", "completed", "failed"]
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "created_at": {
            "type": "integer",
            "description": "Unix 时间戳。"
          }
        },
        "required": [
          "id",
          "task_id",
          "object",
          "model",
          "status",
          "progress",
          "created_at"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            },
            "required": ["message", "code"]
          }
        },
        "required": ["error"]
      }
    }
  }
}